Update Facebook.php
Temporary fix required for a bug in the original repo. Since everything is hardcoded and staticly linked, there unfortunately is no other option than to fork the repo and wait for them to merge the PR. After that this fork can be deleted and we can switch back to the original repo.